core/requirements.txt
alembic==1.0.8
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
Babel==2.3.4
builtwith==1.3.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.4.1
cssselect==1.0.3
decorator==4.0.10
Django==2.0.6
django-admin==1.3.2
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-excel-response2==2.0.8
django-filter==2.0.0
django-relativedelta==1.0.4
django-six==1.0.4
djangorestframework==3.8.2
docutils==0.12
ebaysdk==2.1.5
feedparser==5.2.1
Flask==1.0.2
flask-blueprint==1.2.7
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
get==2018.11.19
gevent==1.3.7
greenlet==0.4.15
httpie==1.0.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
odoo==12.0
parsel==1.5.1
passlib==1.7.1
Pillow==5.3.0
post==2018.11.20
psutil==5.4.8
psycopg2==2.7.6.1
public==2018.11.20
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.0
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==18.0.0
PyPDF2==1.26.0
python-dateutil==2.7.5
python-editor==1.0.4
pytz==2016.7
pywin32==224
query-string==2018.11.20
queuelib==1.5.0
request==2018.11.20
requests==2.20.0
Scrapy==1.5.1
screen==1.0.1
selenium==3.141.0
service-identity==17.0.0
six==1.11.0
source==1.2.0
SQLAlchemy==1.2.14
Twisted==18.9.0
urllib3==1.24.1
virtualenv==16.1.0
w3lib==1.19.0
Werkzeug==0.14.1
xlwt==1.3.0
zope.interface==4.6.0
